float not int

+++ x3d/casaJS.js
@@ -1,4 +1,4 @@
 function forward(){
-	var speed = parseInt(document.getElementById("viewpoint").getAttribute("fieldofview"))+0.1
+	var speed = parseFloat(document.getElementById("viewpoint").getAttribute("fieldofview"))+0.1
 	document.getElementById("viewpoint").setAttribute("fieldofview", speed.toString()) ;
 }
